Moved simplify_num to l10n
