+++ lib/kontext.py
@@ -37,16 +37,6 @@ import fallback_corpus
 from argmapping import ConcArgsMapping, Parameter
 
 
-def simplify_num(v):
-    if v >= 1e9:
-        return '%dG' % (round(v / 1e9, 0))
-    if v >= 1e6:
-        return '%dM' % (round(v / 1e6, 0))
-    if v >= 1e3:
-        return '%dK' % (round(v / 1e3, 0))
-    return '%d' % (round(v / 1e2, 0) * 100,)
-
-
 def join_params(*args):
     """
     This is a convenience function used by HTML templates.
